第一次重构
