+++ .gitignore
@@ -20,4 +20,5 @@
 # Go workspace file
 go.work
 
-.idea # balabala...
+# balabala...
+.idea
